backend/requirements.txt
fastapi
uvicorn[standard]
ytmusicapi
yt-dlp
music-tag
requests
SQLAlchemy
pydantic
alembic
lrclibapi
python-magic
python-dotenv>=1.0.0
Pillow
celery[redis]
redis
python-multipart
typing-extensions
httpx

# Authentication
bcrypt==4.1.2
PyJWT==2.8.0

# Email validation (for Pydantic EmailStr)
email-validator==2.1.0